k12

Module custom

Customized variants.

Modules

borrow 🔒

owned 🔒

Structs

CustomKt

Customized KangarooTwelve hasher generic over rate with owned customization string.

CustomRefKt

Customized KangarooTwelve hasher generic over rate with borrrowed customization string.

Type Aliases

CustomKt128

Customized KT128 hasher with owned customization string.

CustomKt256

Customized KT256 hasher with owned customization string.

CustomRefKt128

Customized KT128 hasher with borrowed customization string.

CustomRefKt256

Customized KT256 hasher with borrowed customization string.