k12::node_turbo_shake

Function parallel

pub(crate) fn parallel<ParSize: ArraySize, const RATE: usize>(
    par_p1600: fn(&mut Array<State1600, ParSize>),
    data: &[u8],
    par_cv_dst: &mut [u8],
)

Parallel version of TurboSHAKE specialized for computation of chaining values.

Ideally, this function should have the following signature:

fn par_turbo_shake<B: Backend, const RATE: usize>(
    p1600: ParFn1600<B>,
    data: &[u8; CHUNK_SIZE * B::PAR_SIZE_1600],
) -> [[u64; {200 - RATE} / 8]; B::PAR_SIZE_1600] { ... }

But it requires advanced const generics or to deal with annoying typenum-based trait bounds, so instead we use “runtime” asserts which should be optimized out by the compiler, see: https://rust.godbolt.org/z/4Y7ervTd7